Finish SARIF reporter and tests

+++ OpalOPC/Controller/Plugin/CommonCredentialsPlugin.cs
@@ -57,7 +57,7 @@ namespace Plugin
             {
                 IEnumerable<string> credpairs = validCredentials.Select(c => $"{c.username}:{c.password}");
                 s_issueTitle = $"Common credentials in use ({string.Join(", ", credpairs)})";
-                return (new CredentialsIssue((int)s_pluginId, s_issueTitle, s_severity, validCredentials), sessions);
+                return (new CredentialsIssue(s_pluginId, s_issueTitle, s_severity, validCredentials), sessions);
             }
 
             return (null, sessions);
